# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
      "ArtifactEntry": {
        "properties": {
          "is_dir": {
            "title": "Is Dir",
            "type": "boolean"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size"
          }
        },
        "required": [
          "path",
          "is_dir"
        ],
        "title": "ArtifactEntry",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MetricSummary": {
        "properties": {
          "key": {
            "title": "Key",
            "type": "string"
          },
          "step": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Step"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          },
          "value": {
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "key",
          "value"
        ],
        "title": "MetricSummary",
        "type": "object"
      },
...
      "RunDetail": {
        "properties": {
          "artifact_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Uri"
          },
          "artifacts": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ArtifactEntry"
            },
            "title": "Artifacts",
            "type": "array"
          },
          "config": {
            "additionalProperties": true,
            "description": "Logged params unflattened back into the nested config tree",
            "title": "Config",
            "type": "object"
          },
          "config_flat": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "The raw flattened params exactly as MLflow stores them",
            "title": "Config Flat",
            "type": "object"
          },
          "config_unflatten_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Set when params could not be unflattened (colliding dotted keys); config falls back to empty",
            "title": "Config Unflatten Error"
          },
          "duration_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Computed from start/end time; null while running",
            "title": "Duration S"
          },
          "end_time": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Unix epoch milliseconds",
            "title": "End Time"
          },
          "experiment_id": {
            "title": "Experiment Id",
            "type": "string"
          },
          "experiment_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experiment Name"
          },
          "final_loss": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Loss"
          },
          "loss_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Which metric final_loss was read from",
            "title": "Loss Key"
          },
          "manifest": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Parsed manifest.json when the run logged one (ergodicio/tsadar#116)",
            "title": "Manifest"
          },
          "metrics": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/MetricSummary"
            },
            "title": "Metrics",
            "type": "array"
          },
          "mlflow_run_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Deep link to this run in the MLflow UI",
            "title": "Mlflow Run Url"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "run_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Name"
          },
          "shot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Shot number, from the data.shotnum param",
            "title": "Shot"
          },
          "spectype": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Spectrum type as logged ('temporal', 'imaging', 'angular', 'angular_full'). A HINT, NOT GROUND TRUTH: log_mlflow runs before fitter.fit, and loadData overwrites spectype from the data file during prepare, so a deck saying 'temporal' run against angular data logs 'temporal'. Confirm from artifact shape (binary/fit_and_data.nc is angular; ele_/ion_fit_and_data.nc is 1D) before rendering anything axis-dependent.",
            "title": "Spectype"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "tsadar's own progress tag ('preprocessing', 'minimizing', ... 'completed'), when logged",
            "title": "Stage"
          },
          "start_time": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Unix epoch milliseconds",
            "title": "Start Time"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "MLflow lifecycle status: RUNNING/FINISHED/FAILED/KILLED",
            "title": "Status"
          },
          "tags": {
            "additionalProperties": {
              "type": "string"
            },
            "default": {},
            "title": "Tags",
            "type": "object"
          },
          "user": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User"
          }
        },
        "required": [
          "run_id",
          "experiment_id"
        ],
        "title": "RunDetail",
        "type": "object"
      },
      "RunPage": {
        "properties": {
          "next_page_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Opaque cursor for the next page. MLflow paginates by token, not offset; null means last page.",
            "title": "Next Page Token"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            },
            "title": "Runs",
            "type": "array"
          }
        },
        "required": [
          "runs",
          "page_size"
        ],
        "title": "RunPage",
        "type": "object"
      },
      "RunSummary": {
        "description": "One row of the run browser table.",
        "properties": {
          "duration_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Computed from start/end time; null while running",
            "title": "Duration S"
          },
          "end_time": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Unix epoch milliseconds",
            "title": "End Time"
          },
          "experiment_id": {
            "title": "Experiment Id",
            "type": "string"
          },
          "experiment_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experiment Name"
          },
          "final_loss": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Loss"
          },
          "loss_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Which metric final_loss was read from",
            "title": "Loss Key"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "run_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Name"
          },
          "shot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Shot number, from the data.shotnum param",
            "title": "Shot"
          },
          "spectype": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Spectrum type as logged ('temporal', 'imaging', 'angular', 'angular_full'). A HINT, NOT GROUND TRUTH: log_mlflow runs before fitter.fit, and loadData overwrites spectype from the data file during prepare, so a deck saying 'temporal' run against angular data logs 'temporal'. Confirm from artifact shape (binary/fit_and_data.nc is angular; ele_/ion_fit_and_data.nc is 1D) before rendering anything axis-dependent.",
            "title": "Spectype"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "tsadar's own progress tag ('preprocessing', 'minimizing', ... 'completed'), when logged",
            "title": "Stage"
          },
          "start_time": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Unix epoch milliseconds",
            "title": "Start Time"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "MLflow lifecycle status: RUNNING/FINISHED/FAILED/KILLED",
            "title": "Status"
          },
          "user": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User"
          }
        },
        "required": [
          "run_id",
          "experiment_id"
        ],
        "title": "RunSummary",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "\nRead layer over the MLflow tracking server for browsing Thomson scattering\nanalysis runs. MLflow remains the source of truth; this service holds no\ndatabase of its own and never writes to it.\n",
    "title": "TSADAR analysis browser API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/runs": {
      "get": {
        "operationId": "list_runs_api_runs_get",
        "parameters": [
          {
            "description": "Experiment name (or id)",
            "in": "query",
            "name": "experiment",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Experiment name (or id)",
              "title": "Experiment"
            }
          },
          {
            "description": "Shot number, matched against params.\"data.shotnum\"",
            "in": "query",
            "name": "shot",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Shot number, matched against params.\"data.shotnum\"",
              "title": "Shot"
            }
          },
          {
            "description": "MLflow lifecycle status: RUNNING, FINISHED, FAILED, KILLED",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "MLflow lifecycle status: RUNNING, FINISHED, FAILED, KILLED",
              "title": "Status"
            }
          },
          {
            "description": "tsadar progress tag, e.g. 'minimizing', 'completed'",
            "in": "query",
            "name": "stage",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "tsadar progress tag, e.g. 'minimizing', 'completed'",
              "title": "Stage"
            }
          },
          {
            "description": "Submitting user (mlflow.user tag)",
            "in": "query",
            "name": "user",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Submitting user (mlflow.user tag)",
              "title": "User"
            }
          },
          {
            "description": "Free-text substring match on run name. '%' and '_' act as SQL LIKE wildcards (any-sequence and any-character) rather than literals; MLflow's filter grammar has no ESCAPE clause to make them literal.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Free-text substring match on run name. '%' and '_' act as SQL LIKE wildcards (any-sequence and any-character) rather than literals; MLflow's filter grammar has no ESCAPE clause to make them literal.",
              "title": "Q"
            }
          },
          {
            "description": "Sort key; prefix with '-' for descending. One of created, name, status, shot, loss. 'loss' orders on the 'overall loss' metric specifically -- unlike final_loss it cannot fall back to 'min loss'/'epoch loss', so runs lacking 'overall loss' sort as null; check loss_key when displaying the value. Duration is not sortable: it is computed from timestamps and MLflow cannot order by it.",
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Sort key; prefix with '-' for descending. One of created, name, status, shot, loss. 'loss' orders on the 'overall loss' metric specifically -- unlike final_loss it cannot fall back to 'min loss'/'epoch loss', so runs lacking 'overall loss' sort as null; check loss_key when displaying the value. Duration is not sortable: it is computed from timestamps and MLflow cannot order by it.",
              "title": "Sort"
            }
          },
          {
            "description": "Rows per page",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Rows per page",
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "description": "Cursor from a previous response's next_page_token",
            "in": "query",
            "name": "page_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Cursor from a previous response's next_page_token",
              "title": "Page Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search runs",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}": {
      "get": {
        "operationId": "get_run_api_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run detail with config tree and artifact listing",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}/artifacts/{artifact_path}": {
      "get": {
        "description": "Serve artifact bytes through the API so the frontend needs no S3 credentials.",
        "operationId": "get_artifact_api_runs__run_id__artifacts__artifact_path__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "artifact_path",
            "required": true,
            "schema": {
              "title": "Artifact Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {}
            },
            "description": "Artifact bytes"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream an artifact",
        "tags": [
          "runs"
        ]
      }
    },
...
  }
}